refactor : [로또] 패키지 구조 변경
