+++ src/main/java/lotto/domain/Lotto.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import lotto.pattern.LottoNumberGenerateStrategy;
