+++ src/main/java/lotto/domain/LottoGroup.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import lotto.pattern.LottoNumberGenerateStrategy;
