+++ src/main/java/lotto/domain/Rank.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import java.util.Map;
 import java.util.Optional;
