+++ src/main/java/lotto/util/SplitString.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.util;
 
 import java.util.List;
 import java.util.stream.Collectors;
