+++ src/test/java/lotto/domain/LottoTotalPriceTest.java
@@ -1,5 +1,6 @@
-package lotto;
+package lotto.domain;
 
+import lotto.domain.LottoTotalPrice;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
