+++ src/test/java/lotto/util/SplitStringTest.java
@@ -1,5 +1,6 @@
-package lotto;
+package lotto.util;
 
+import lotto.util.SplitString;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
